+++ src/main/java/eu/evops/maven/plugins/cucumber/parallel/Run.java
@@ -9,7 +9,6 @@ import org.apache.commons.io.filefilter.DirectoryFileFilter;
 import org.apache.commons.io.filefilter.NameFileFilter;
 import org.apache.maven.artifact.Artifact;
 import org.apache.maven.plugin.AbstractMojo;
-import org.apache.maven.plugin.MojoExecutionException;
 import org.apache.maven.plugin.MojoFailureException;
 import org.apache.maven.plugins.annotations.Mojo;
 import org.apache.maven.plugins.annotations.Parameter;
@@ -120,7 +119,7 @@ public class Run extends AbstractMojo {
     /**
      * Treat undefined and pending steps as errors.
      */
-    @Parameter
+    @Parameter(property = "cucumberRunner.strict")
     boolean strict = true;
 
     /**
@@ -146,7 +145,7 @@ public class Run extends AbstractMojo {
     private File threadFolder;
     private String streamingFormatterClassName = StreamingJSONFormatter.class.getName();
 
-    public void execute() throws MojoExecutionException, MojoFailureException {
+    public void execute() throws MojoFailureException {
         setThreadCount();
         setOutputFolder();